Updates
The AccompanyRhythmController.cs file keeps the rhythm of the object in accordance with the BPM.

The attack01.json file is an example json file.

The BulletSpawner.cs file spawns objects according to json. Also, the json reading process is in this script.

The BulletSpawnerEditor.cs file adds a button to the inspector.